move Relational solving to _solveset
Since Piecewise can contain Relationals we need to either
call solveset from _solveset or move the Relational solving
blocks into _solveset. The latter has been done.

When solving Piecewise, the conditions can only be solved
in the Real domain so the domain intersection with S.Reals
is passed to the as_expr_set_pairs function. The domain
itself is not changed, however, so if an attempt to solve
a Piecewise with expressions that are Relational is made an
error will be raised just like it would be raised if the
Relational didn't appear in a Piecewise:

>>> solveset(Lt(x, 2), x)
...
NotImplementedError:
Inequalities in the complex domain are not supported. Try the real
domain by setting domain=S.Reals